% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{calcInitChange}
\alias{calcInitChange}
\title{Calculate changes to concepts being varied in scenario}
\usage{
calcInitChange(V_Cn, TargetV_Cn, RemIncr, Type = "Exponential")
}
\arguments{
\item{V_Cn}{a numeric vector identifying the values of concepts}

\item{TargetV_Cn}{a numeric vector identifying the target values of concepts
for the scenario.}

\item{RemIncr}{a number identifying the remaining number of increments to
complete the scenario.}

\item{Type}{a string identifying the growth function. May be either
'Exponential' or 'Linear'.}
}
\value{
a list having 2 named components. Ratio is the ratio of change of the
'independent' concepts. Values is a numeric vector of the value of concepts
after 'independent' concepts have been incremented.
}
\description{
\code{calcInitChange} calculates changes to concepts being varied in a
scenario
}
\details{
In a scenario, the values of one or a few concepts are changed and then the
model is run to calculate a new equilibrium state of concept values. The
changes to these 'independent' concepts are made in small increments to model
assumptions about how the values will change over time. In each increment,
the change for the increment is calculated from the concept values of the
previous increment, the scenario target concept values, and the number of
remaining increments. Calculation of the increments assumes that change will
occur at a constant exponential rate over the course of the remaining
increments.
}
