% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{conformScenario}
\alias{conformScenario}
\title{Modify a scenario to conform with a changed model}
\usage{
conformScenario(ConceptVars_, ScenarioValues_df)
}
\arguments{
\item{ConceptVars_}{a string vector of the variable names of model concepts}

\item{ScenarioValues_df}{a data frame of scenario concept values}
}
\value{
a data frame of scenario concept values that is consistent with the
model concepts
}
\description{
\code{conformScenario} modifies an existing scenario definition so that it
conforms to the current model definition.
}
\details{
This function modifies a scenario definition so that it conforms to the
the current model definition. It does this my removing entries for concepts
that are not part of the model and by adding "blank" entries for concepts
that are in the model but not in the scenario. The function is designed to be
called by the 'initializeCopyScenario' function and the 'loadScenario'
function. Note that the function does not check whether the values for
concepts are consistent with the concept definition. The user needs to check
consistency of values and validation will check for inconsistencies as well.
}
