% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{copyDemoModels}
\alias{copyDemoModels}
\title{Copy demo model files from FSDM package}
\usage{
copyDemoModels(ToFolder)
}
\arguments{
\item{ToFolder}{a string identifying the path to the folder on the users
computer that the demo files are to be copied to.}
}
\description{
\code{copyDemoModels} copies demo FSDM model files to specified folder
}
\details{
This function copies demo FSDM model files from the FSDM package to a
specified folder. If a folder named 'models' is located in the specified
folder, the function copies the demo models to that folder. If the folder
does not exist, then it is created and the demo models copied to it.
}
