% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{getEffects}
\alias{getEffects}
\title{Get data on effect relationships of a causal concept}
\usage{
getEffects(Model_ls, ConceptName)
}
\arguments{
\item{Model_ls}{a list that includes components for model concepts and
relations}

\item{ConceptName}{a string containing the name of the causal concept}
}
\value{
a data frame containing the names, directions, magnitudes, and
descriptions of all the effects or NULL if there are no effects.
}
\description{
\code{getEffects} extracts data on all the effect relationships of an
identified causal concept.
}
\details{
The function extracts the data on the effect relationships of an identified
causal concept. This includes the names of the affected concepts and the
respective effect directions, magnitudes, and descriptions.
}
