% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{idScenWithOutputs}
\alias{idScenWithOutputs}
\title{Identify scenarios that have model outputs}
\usage{
idScenWithOutputs(ModelsDir, ModelName)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which}

\item{ModelName}{a string representation of the model name.}
}
\value{
a string vector of the names of scenarios having model outputs.
}
\description{
\code{idScenWithOutputs} returns the names of scenarios that have model run
outputs.
}
\details{
This function takes the name of a model and returns a vector of the names
of all the model scenarios that have model outputs.
}
