% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{initializeNewModel}
\alias{initializeNewModel}
\title{Initialize a new model.}
\usage{
initializeNewModel(ModelsDir, ModelName, Author)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which
the model is located.}

\item{ModelName}{a string identifying the model name.}

\item{Author}{a string identifying the author's name.}
}
\value{
a list containing values for name, parent, created, and lastedit.
}
\description{
\code{initializeNewModel} initializes a new model by creating a directory
with the model name. Creates and saves a model status list.
}
\details{
This function initializes a new model with the given model name. It does this
by creating a directory with the model name and a list to store the model
status. This list is saved in the model directory in JSON format and is
also returned by the function. The status list contains the name of the
model, the parent (none), the date and time is was created, and the date
and time it was last edited (same as creation time).
}
