% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{initializeNewScenario}
\alias{initializeNewScenario}
\title{Initialize a new scenario.}
\usage{
initializeNewScenario(ModelsDir, ModelName, ScenarioName, Concepts_df,
  NumIncr)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which}

\item{ModelName}{a string representation of the model name.}

\item{ScenarioName}{a string representation of the scenario name.}

\item{Concepts_df}{a data frame containing the concept data.}

\item{NumIncr}{an integer identifying the number of growth increments for
the scenario.}
}
\value{
a list containing a status list and a scenario values data frame.
}
\description{
\code{initializeNewScenario} initializes a new scenario by creating a
directory with the scenario name and returns a list containing scenario
status, scenario values, and number of growth increments components.
}
\details{
This function initializes a new scenario with the given scenario name. It
does this by creating a directory with the scenario name and a data frame
to store the scenario data. This list is returned and the components are
used in the scenario reactive object. The components are also saved in the
scenario directory in JSON format.
}
