% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{initRelationsEntry}
\alias{initRelationsEntry}
\title{Initialize a initial relations entry for a new concept}
\usage{
initRelationsEntry(VarName)
}
\arguments{
\item{VarName}{a string representation of the concept variable name}
}
\value{
a data frame which includes all the mandatory relations fields with
the name field populated with the concept variable name and all other fields
populated with empty fields
}
\description{
\code{initRelationsEntry} creates a initial relations entry for concept
that is being added to the model
}
\details{
This function creates an initial relations entry data for a concept. This
entry data has the concept variable name and empty fields for all other data
items. The server script calls this function when a concept is created and
adds the resulting entry to the relations table.
}
