% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{listScenarios}
\alias{listScenarios}
\title{Creates a list of scenarios and whether validated.}
\usage{
listScenarios(ModelsDir, ModelName)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which
the operating model is located.}

\item{ModelName}{the name of the model}
}
\value{
a list having four components, a vector of the names of scenarios
that were properly validated, a vector of the names of scenarios that were
not validated, a vector of the names of all scenarios, and a vector
of the names of scenarios that have outputs.
}
\description{
\code{listScenarios} creates a list of scenarios and identifies which have
been validated.
}
\details{
This function creates a list of scenarios for a model and checks whether each
scenario has been validated. Scenarios are identified as being validated if
they were successfully validated and if the validation time stamp is later
than the 'lastedit' time stamp for the model and the 'lastedit' time stamp
for the scenario. The returned list has two components. The first is a vector
containing the names that have been successfully validated. The second is a
vector of the names that have not been validated; either because validation
was not successful or because changes were made to the model or the scenario
after validation was done.
}
