% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{loadModelConcepts}
\alias{loadModelConcepts}
\title{Load the concept file for a model.}
\usage{
loadModelConcepts(ModelsDir, ModelName)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which}

\item{ModelName}{a string representation of the model name.}
}
\value{
a data frame containing the model concept information.
}
\description{
\code{loadModelConcepts} reads the file that contains model concept information
and returns a data frame representation.
}
\details{
This function reads the model concept file for a specified model and returns
a data frame containing the information.
}
