% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{loadModelRelations}
\alias{loadModelRelations}
\title{Load the relations file for a model.}
\usage{
loadModelRelations(ModelName)
}
\arguments{
\item{ModelName}{a string representation of the model name.}
}
\value{
a data frame containing the model relations information.
}
\description{
\code{loadModelRelations} reads the file that contains model relations
information and returns a data frame representation.
}
\details{
This function reads the model relations file for a specified model and
returns a data frame containing the information.
}
