% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{loadModelStatus}
\alias{loadModelStatus}
\title{Load the model status file for a model.}
\usage{
loadModelStatus(ModelName, Author = NULL)
}
\arguments{
\item{ModelName}{a string identifying the model name.}

\item{Author}{a string identifying the author's name.}
}
\value{
a model status list.
}
\description{
\code{loadModelStatus} reads a model status file and returns a list
containing the status information.
}
\details{
This function reads the model status JSON file for a specified model and
creates a list containing the model status information.
}
