% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{loadScenario}
\alias{loadScenario}
\title{Loads the files for a scenario.}
\usage{
loadScenario(ModelName, ConceptVars_, ScenarioFileName)
}
\arguments{
\item{ModelName}{a string representation of the model name.}

\item{ConceptVars_}{a string vector of the model concept variable names.}

\item{ScenarioName}{a string representation of the scenario name.}
}
\value{
a list containing a status list and a scenario values data frame.
}
\description{
\code{loadScenario} reads the files that contain scenario information
and returns a list containing scenario status and values information.
}
\details{
This function reads the scenario status file and scenario file for a
specified scenario and returns a list whose components are a list containing
the scenario status information and a data frame containing the scenario
values information.
}
