% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{makeAdjacencyMatrix}
\alias{makeAdjacencyMatrix}
\title{Creates an adjacency matrix from an FSDM relations list}
\usage{
makeAdjacencyMatrix(Relations_ls, Type = "Logical")
}
\arguments{
\item{Relations_ls}{a list of FSDM relations}

\item{Type}{a string identifying the type of matrix to create. "Logical"
produces a matrix where the existence of a relationship is noted with TRUE
and FALSE. "Values" produces a list of two matrices where one matrix
identifies the causal weitht of each relationship and a second matrix
identifies the causal direction of each relationship.}
}
\value{
a matrix of logical values or string values
}
\description{
\code{makeAdjacencyMatrix} creates an adjacency matrix from an FSDM relations
list
}
\details{
This function creates an adjacency matrix from an FSDM relations list. The
adjacency matrix is a square matrix with as many rows and columns as the
number of concepts in the model. The rows and columns are named with the
concept variable names. The rows represent the causal side of the
relationship and the columns represent the affected side. The values in the
matrix are logicals with TRUE meaning that a relationship exists and FALSE
meaning that it does not.
}
