% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{makeDot}
\alias{makeDot}
\title{Create a DOT file for GraphViz}
\usage{
makeDot(Relations_ls, Concepts_df, RowGroup = "All", ColGroup = "All",
  orientation = "Portrait", rankdir = "Top-to-Bottom", shape = "box",
  Show = "label")
}
\arguments{
\item{Relations_ls}{a list of model relations.}

\item{Concepts_df}{a data frame with model concepts.}

\item{RowGroup}{a string identifying the names of the name of the group that
selects rows from the relationship table to plot. The default 'All' selects
all the rows.}

\item{ColGroup}{a string identifying the name of the group that selects
columns from the relationship table to plot. The default 'All' selects all
the columns.}

\item{orientation}{a string identifying the GraphViz layout orientation
('Portrait' or 'Landscape').}

\item{rankdir}{a string identifying the graph orientation:
'TB' for top to bottom, 'LR' for left to right.}

\item{shape}{a string identifying the shape of the graph nodes (e.g. 'box').}

\item{Show}{a string identifying how to label the edges. The default value
"label" results in showing the fuzzy label (e.g. VL, L, M, H, VH). The
alternative, "value", results in showing the equivalent numeric value.}
}
\value{
A string specification of a DOT.
}
\description{
\code{makeDotFile} create and save a dot file to be displayed by GraphViz
}
\details{
This function writes out a dot file to be rendered using GraphViz.
}
