% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{mapRelations}
\alias{mapRelations}
\title{Plot relationships}
\usage{
mapRelations(Model_ls, FromConcept = NULL, FromGroup = "All",
  ToGroup = "All")
}
\arguments{
\item{Model_ls}{a list that includes components for model concepts and
relations}

\item{FromConcept}{the name of a selected causal concept or NULL}

\item{FromGroup}{the name of the group of the causal concepts to display or
NULL}

\item{ToGroup}{the name of the group of the receiving concepts to display or
NULL}
}
\value{
a logical value identifying whether the plot can be created
}
\description{
\code{mapRelations} displays a plot of concept relations, highlighting the
relations from or to a highlighted concept.
}
\details{
The function maps relationships between concepts by plotting concepts in
two parallel vertical lines and then showing relationships between concepts
by drawing lines between the related concepts. The selected concept and
related concepts are placed at the top of the plot and the lines are
highlighted.
}
