% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{rescale}
\alias{rescale}
\title{Rescale value}
\usage{
rescale(Value, FromRange, ToRange)
}
\arguments{
\item{Value}{a numeric value to be rescaled from the input range to the
output range.}

\item{FromRange}{a numeric vector of length 2 in which the first value is
the minimum value in the range and the second value is the maximum value in
the range.}

\item{ToRange}{a numeric vector of length 2 in which the first value is the
minimum value in the range and the second value is the maximum value in the
range.}
}
\value{
A numeric value in the output range.
}
\description{
\code{rescale} rescales a value from a specified input range to a specified
output range.
}
\details{
This function rescales a value from a specified input range to a specified
output range. The default output range is 0 to 100 because that is the range
used to represent concepts in FSDM models.
}
