% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{runFuzzyModel}
\alias{runFuzzyModel}
\title{Run the fuzzy model}
\usage{
runFuzzyModel(M, S, Type, NumIncr = 100)
}
\arguments{
\item{M}{a fuzzy model created using the create FuzzyModel function.}

\item{S}{a scenario created using the loadScenario function.}

\item{Type}{a string that may have either the value 'Linear' or 'Exponential'
which identifies whether independent concepts should change linearly or
exponentially.}

\item{NumIncr}{a numeric value identifying the number of increments to use
for building up the starting changes.}

\item{MaxIter}{a numeric value identifying the maximum number of iterations
for the inner loop.}

\item{OpRange}{a 2-element vector identifying the operating range of the
model. The default values are c(0.01, 99.99), just short of 0 and 100 to
avoid values that either result in no change or infinite change.}
}
\value{
A list containing 3 components:
Summary - the final relative concept values (e.g. 0 - 100) for each concept
and increment;
ScaleSummary - the final concept values in the nominal measurement units for
each concept and increment;
Full - the relative concept values (e.g. 0 - 100) for each increment and
each iteration.
}
\description{
\code{runFuzzyModel} runs a fuzzy model given scenario inputs for starting
values and starting changes.
}
\details{
This function runs a FCM model that was created using the createFuzzyModel
function with a scenarios that was created using the createFuzzyScenario
function. The function uses a nested loop to calculate final values. The
outer loop applies the starting changes in small increments so that the
changes appropriate for applying 'point elasticities'. The inner loop
iterates the model until a stable equilibrium result is achieved for each
increment.
}
