% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{saveModel}
\alias{saveModel}
\title{Saves all the model components as JSON files.}
\usage{
saveModel(ModelsDir, ModelData)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which}

\item{ModelData}{a model.}
}
\value{
no return value. Has side effect of saving the model status list,
model concepts data frame, and model relations data frame as JSON-formatted
files.
}
\description{
\code{saveModel} saves the model status, model concepts, and model relations
as JSON files.
}
\details{
Models are composed of 3 objects: a model status list, a model concepts
data frame, and a model relations data frame. This function saves these
objects as JSON-formatted files.
}
