% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FSDM.R
\name{saveScenario}
\alias{saveScenario}
\title{Saves all the scenario components as JSON files.}
\usage{
saveScenario(ScenarioData)
}
\arguments{
\item{ScenarioData}{a list having status and values components.}
}
\value{
no return value. Has side effect of saving the scenario status list
and values data frame.
}
\description{
\code{saveScenario} saves the scenario status and values as JSON files.
}
\details{
This function saves the scenario status and values as JSON files in the
scenario directory for the model. The status information is saved in the
status.json file and the scenario values are saved in the scenarios.json
file.
}
