% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{validateScenario}
\alias{validateScenario}
\title{Validates scenario with model.}
\usage{
validateScenario(Values_df, Concepts_df)
}
\arguments{
\item{Values_df}{a data frame containing the scenario values}

\item{Concepts_df}{a data frame containing the model concepts}
}
\value{
a list having three components, a logical identifying whether the
scenario validates, a list of all validation errors, and a character string
with the time stamp.
}
\description{
\code{validateScenario} validates the scenario with the model.
}
\details{
This function compares the values for the scenario with the model and
determines whether the scenario has values for all concepts, whether the
values are consistent with the value ranges for the concepts, and whether
the starting changes are consistent with the ranges and don't vary too many.
}
