% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{initializeCopyScenario}
\alias{initializeCopyScenario}
\title{Initialize a new scenario by copying an existing scenario.}
\usage{
initializeCopyScenario(ModelsDir, ModelName, ConceptVars_, ScenarioName,
  CopyScenarioName, NumIncr)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which
the model is located.}

\item{ModelName}{a string representation of the model name.}

\item{ConceptVars_}{a string vector of the model concept variable names.}

\item{ScenarioName}{a string representation of the new scenario name.}

\item{CopyScenarioName}{a string representation of the name of the scenario
to copy.}

\item{NumIncr}{an integer identifying the number of growth increments for the
scenario.}
}
\value{
a list containing a status list and a scenario values data frame.
}
\description{
\code{initializeCopyScenario} initializes a new scenario by creating a
directory with the scenario name and copying the contents of an existing
scenario into that directory.
}
\details{
This function initializes a new scenario with the given scenario name from an
existing scenario. It does this by creating a directory with the scenario
name and copying the scenario files for an existing scenario into it. It
creates a new scenario status list which identifies the name of the new
scenario and the name of the parent scenario it is a copy of. It also
identifies the number of growth increments in the new scenario.
}
