% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{loadModelStatus}
\alias{loadModelStatus}
\title{Load the model status file for a model.}
\usage{
loadModelStatus(ModelsDir, ModelName, Author = NULL)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which}

\item{ModelName}{a string representation of the model name.}

\item{Author}{a string identifying the author's name}
}
\value{
a list containing values for name, parent, created, and lastedit.
}
\description{
\code{loadModelStatus} reads a model status file and returns a list
containing the status information.
}
\details{
This function reads the model status JSON file for a specified model and
creates a list containing the model status information.
}
