% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{loadScenario}
\alias{loadScenario}
\title{Loads the files for a scenario.}
\usage{
loadScenario(ModelsDir, ModelName, ConceptVars_, ScenarioFileName)
}
\arguments{
\item{ModelsDir}{a string identifying the path to the models folder in which}

\item{ModelName}{a string representation of the model name.}

\item{ConceptVars_}{a string vector of the model concept variable names.}

\item{ScenarioFileName}{a string representation of the scenario name.}
}
\value{
a list containing a status list, a scenario values data frame, and
the number of growth increments for the scenario.
}
\description{
\code{loadScenario} reads the files that contain scenario information
and returns a list containing scenario status, values, and increments
information.
}
\details{
This function reads the scenario status file and scenario file for a
specified scenario and returns a list whose components are a list containing
the scenario status information, a data frame containing the scenario
values information, and an integer identifying the number of growth
increments.
}
