% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsdm.R
\name{runFuzzyModel}
\alias{runFuzzyModel}
\title{Run the fuzzy model}
\usage{
runFuzzyModel(M, S, Type)
}
\arguments{
\item{M}{a fuzzy model created using the create FuzzyModel function.}

\item{S}{a scenario created using the loadScenario function.}

\item{Type}{a string that may have either the value 'Linear' or 'Exponential'
which identifies whether independent concepts should change linearly or
exponentially.}
}
\value{
A list containing 3 components:
Summary - a matrix containing the final relative concept values (e.g.
0 - 100) for each concept and increment;
ScaleSummary - a matrix containing the final concept values in the nominal
measurement units for each concept and increment;
Full - the relative concept values (e.g. 0 - 100) for each increment and
each iteration.
Message - a string containing a final completion message identifying whether
the model run completed successfully and some information about the run.
Success - a logical vector identifying whether the solution converged on each
increment.
NumIter - a numeric vector identifying the number of iterations to model
convergence for each increment.
}
\description{
\code{runFuzzyModel} runs a fuzzy model given scenario inputs for starting
values and starting changes.
}
\details{
This function runs a FCM model that was created using the createFuzzyModel
function with a scenarios that was created using the createFuzzyScenario
function. The function uses a nested loop to calculate final values. The
outer loop applies the starting changes in small increments so that the
changes appropriate for applying 'point elasticities'. The inner loop
iterates the model until a stable equilibrium result is achieved for each
increment.
}
